--- lib/utils/stepper.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Stepper extends StatefulWidget {
-  const Stepper({super.key});
-
-  @override
-  State<Stepper> createState() => _StepperState();
-}
-
-class _StepperState extends State<Stepper> {
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
